requirements.txt
Django
djangorestframework
django-cors-headers
jazzmin
python-dotenv
djangorestframework-simplejwt
reportlab
python-dateutil
